+++ webapp/tests/integration/health-api.test.js
@@ -3,8 +3,8 @@
  * @tagline         Integration tests for Health API endpoints
  * @description     Integration tests for health and metrics API endpoints
  * @file            webapp/tests/integration/health-api.test.js
- * @version         1.4.1
- * @release         2025-12-31
+ * @version         1.4.2
+ * @release         2026-01-01
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/