+++ webapp/tests/setup/env-setup.js
@@ -3,8 +3,8 @@
  * @tagline         Jest Environment Setup
  * @description     Sets up environment variables for Jest tests
  * @file            webapp/tests/setup/env-setup.js
- * @version         1.4.1
- * @release         2025-12-31
+ * @version         1.4.2
+ * @release         2026-01-01
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/