+++ webapp/tests/unit/controller/handlebar-logical-helpers.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-114: {{and}}, {{or}}, {{eq}} logical helpers
  * @description     Tests for logical and comparison helpers with subexpression support
  * @file            webapp/tests/unit/controller/handlebar-logical-helpers.test.js
- * @version         1.4.1
- * @release         2025-12-31
+ * @version         1.4.2
+ * @release         2026-01-01
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/