+++ webapp/tests/unit/controller/markdown-ignore.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for markdown controller ignore functionality
  * @description     Tests for .markdown [ignore] section pattern matching and filtering (W-120)
  * @file            webapp/tests/unit/controller/markdown-ignore.test.js
- * @version         1.4.1
- * @release         2025-12-31
+ * @version         1.4.2
+ * @release         2026-01-01
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/